Fix and improve the timing measurements:
The timing measurements are incorrect since the
order of the three calls to the library function
'clock_gettime' in the function 'ntime_clocks'
isn't the reverse of that of the three calls to
'clock_gettime' in the function 'ntime_init' in
'src/clocks-impl.h'.

The measurements improvement results from the
timing adjustments introduced by the function
'clocks_adjust' in 'src/clocks.[hc]'. (See the
comment text preceding the implementation code
of this function.)